Removing reads already assigned.
reads already assigned to blocks are removed from the cluster list. Then a dictionary for the cluster is instantiated with its keys as blocks numbers and the corresponding values will be lists that contain the reads assigned to that block. After all the reads are assigned we are ready to print the cluster.